refactor: merge useFloatingDateOpacity into useFloatingDate
