+++ app/containers/Separator/index.ts
@@ -1,4 +1,3 @@
 export { default as DateSeparator, DateSeparatorLabel } from './DateSeparator';
 export { default as FloatingDateSeparator } from './FloatingDateSeparator';
 export { default as MessageSeparator } from './MessageSeparator';
-export { useFloatingDateOpacity } from './useFloatingDateOpacity';
