+++ app/views/RoomView/List/hooks/useFloatingDate.ts
@@ -1,12 +1,19 @@
 import { useState } from 'react';
 import { type FlatListProps, type ViewToken } from 'react-native';
+import { type SharedValue, useSharedValue, withDelay, withTiming } from 'react-native-reanimated';
 
 import { type TAnyMessageModel } from '../../../../definitions';
+
+const HIDE_DELAY = 1000;
+const FADE_IN_DURATION = 150;
+const FADE_OUT_DURATION = 300;
 
 type TViewabilityConfigCallbackPairs = NonNullable<FlatListProps<TAnyMessageModel>['viewabilityConfigCallbackPairs']>;
 
 interface IUseFloatingDate {
 	ts: Date | string | null;
+	opacity: SharedValue<number>;
+	show: () => void;
 	viewabilityConfigCallbackPairs: TViewabilityConfigCallbackPairs;
 }
 
@@ -26,6 +33,8 @@ export const getTopVisibleTs = (viewableItems: ViewToken<TAnyMessageModel>[]): D
 
 export const useFloatingDate = (): IUseFloatingDate => {
 	const [ts, setTs] = useState<Date | string | null>(null);
+	const opacity = useSharedValue(0);
+	const isFadingIn = useSharedValue(false);
 
 	const [viewabilityConfigCallbackPairs] = useState<TViewabilityConfigCallbackPairs>(() => [
 		{
@@ -39,5 +48,30 @@ export const useFloatingDate = (): IUseFloatingDate => {
 		}
 	]);
 
-	return { ts, viewabilityConfigCallbackPairs };
+	const hide = (): void => {
+		'worklet';
+
+		opacity.set(withDelay(HIDE_DELAY, withTiming(0, { duration: FADE_OUT_DURATION })));
+	};
+
+	const show = (): void => {
+		'worklet';
+
+		if (isFadingIn.get()) {
+			return;
+		}
+		if (opacity.get() === 1) {
+			hide();
+			return;
+		}
+		isFadingIn.set(true);
+		opacity.set(
+			withTiming(1, { duration: FADE_IN_DURATION }, (): void => {
+				isFadingIn.set(false);
+				hide();
+			})
+		);
+	};
+
+	return { ts, opacity, show, viewabilityConfigCallbackPairs };
 };
